+++ src/edb.cpp
@@ -175,8 +175,8 @@ void address_t::normalize() {
 
 namespace v1 {
 
-bool debuggeeIs32Bit() { return pointer_size()==sizeof(std::uint32_t); }
-bool debuggeeIs64Bit() { return pointer_size()==sizeof(std::uint64_t); }
+bool debuggeeIs32Bit() { return pointer_size() == sizeof(std::uint32_t); }
+bool debuggeeIs64Bit() { return pointer_size() == sizeof(std::uint64_t); }
 
 //------------------------------------------------------------------------------
 // Name: set_cpu_selected_address
@@ -1411,7 +1411,7 @@ std::shared_ptr<IBreakpoint> find_triggered_breakpoint(address_t address) {
 // Name: pointer_size
 // Desc:
 //------------------------------------------------------------------------------
-int pointer_size() {
+size_t pointer_size() {
 
 	if(debugger_core) {
 		return debugger_core->pointer_size();
